requirements.txt
attrs==26.1.0
bluesky-navdata==1.0.0
-e git+https://github.com/TUDelft-CNS-ATM/bluesky.git@87b156f6445d5956d22302a594f08c4171bd4ac4#egg=bluesky_simulator
casadi==3.7.2
cdsapi==0.7.7
certifi==2026.2.25
cffi==2.0.0
cfgrib==0.9.15.1
cftime==1.6.5
charset-normalizer==3.4.7
click==8.3.2
-e git+https://github.com/dlr-pa/climaccf.git@018471922764b649ef23285f64919243b68a19b6#egg=climaccf
cloudpickle==3.1.2
contourpy==1.3.2
cycler==0.12.1
dask==2026.3.0
eccodes==2.46.0
eccodeslib==2.46.2.19
eckitlib==2.0.7.19
ecmwf-api-client==1.6.5
ecmwf-datastores-client==0.5.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
filelock==3.28.0
findlibs==0.1.2
flexcache==0.3
flexparser==0.4
fonttools==4.62.1
fsspec==2026.3.0
future==1.0.0
geojson==3.2.0
geojsoncontour==0.5.1
h5netcdf==1.8.1
hyperopt==0.2.7
idna==3.11
importlib_metadata==9.0.0
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
locket==1.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
MetPy==1.6.3
mpmath==1.3.0
msgpack==1.1.2
multiurl==0.3.7
netCDF4==1.7.4
networkx==3.4.2
numpy==2.2.6
openap==2.4
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
partd==1.4.2
pillow==12.2.0
Pint==0.24.4
platformdirs==4.9.6
pluggy==1.6.0
pooch==1.9.0
py4j==0.10.9.9
pycontrails==0.57.0
pycparser==3.0
Pygments==2.20.0
pyparsing==3.3.2
pyproj==3.7.1
pytest==9.0.3
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.33.1
scikit-learn==1.7.2
scipy==1.15.3
setuptools-scm==8.1.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tomli==2.4.1
toolz==1.1.0
torch==2.11.0
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
xarray==2024.10.0
xlrd==2.0.2
zipp==3.23.1
zmq==0.0.0
